refactor(crm): konsolidera API-hjälpare, domänlager och Supabase-klientanvändning
Wave CRM-1: Skapar app/api/crm/_shared.ts med ok, routeError, validationError,
requireCrmUser, requireCrmAdmin och requireSignedInUser. Tar bort identiska kopior
ur alla 9 _lib.ts-filer — varje fil innehåller nu enbart Zod-scheman.

Wave CRM-2: goals/route.ts GET byter getSupabaseAdmin() mot createRouteHandlerClient
(mål-läsning ska följa RLS). Ringlists admin-klientanvändning dokumenteras per route.

Wave CRM-3: lib/crm/ → lib/domains/crm/, 22 importer uppdaterade.

Wave CRM-4: CrmOverview.tsx lokal EmptyState ersatt med components/ui/EmptyState.
