+++ app/api/crm/ringlists/_lib.ts
@@ -1,6 +1,5 @@
-import { NextResponse } from 'next/server';
 import { z } from 'zod';
-import { getCurrentUser } from '@/lib/auth/route';
+export { ok, routeError, validationError, requireCrmAdmin } from '../_shared';
 
 export const assignCrmRinglistProspectsSchema = z.object({
   prospect_ids: z.array(z.string().uuid()).min(1, 'Välj minst ett prospekt'),
@@ -28,43 +27,3 @@ export const importCrmRinglistRowsSchema = z.object({
   })).min(1, 'Minst en rad krävs').max(500, 'Importen är för stor, dela upp filen i mindre delar'),
 });
 
-export function ok<T>(data: T, status = 200) {
-  return NextResponse.json({ ok: true, data }, { status, headers: { 'Cache-Control': 'no-store' } });
-}
-
-export function routeError(status: number, code: string, message: string, details?: unknown) {
-  return NextResponse.json(
-    {
-      ok: false,
-      error: message,
-      errorDetails: { code, message, ...(details !== undefined ? { details } : {}) },
-    },
-    { status, headers: { 'Cache-Control': 'no-store' } },
-  );
-}
-
-export function validationError(parsedError: z.ZodError) {
-  const flattened = parsedError.flatten();
-  const fieldErrorGroups = Object.values(flattened.fieldErrors);
-
-  for (const messages of fieldErrorGroups) {
-    const firstMessage = messages?.find(Boolean);
-    if (firstMessage) return routeError(400, 'validation_error', firstMessage, flattened);
-  }
-
-  return routeError(400, 'validation_error', flattened.formErrors.find(Boolean) || 'Invalid request', flattened);
-}
-
-export async function requireCrmAdmin() {
-  const currentUser = await getCurrentUser();
-
-  if (!currentUser) {
-    return { currentUser: null, response: routeError(401, 'unauthorized', 'Unauthorized') };
-  }
-
-  if (currentUser.role !== 'admin') {
-    return { currentUser: null, response: routeError(403, 'forbidden', 'Forbidden') };
-  }
-
-  return { currentUser, response: null };
-}
